requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.7
django-cors-headers==4.9.0
django-extra-fields==3.0.2
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
et_xmlfile==2.0.0
Faker==37.11.0
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
openpyxl==3.1.5
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-decouple==3.8
python-magic==0.4.27
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
# PDF Generation - using xhtml2pdf (pure Python, no system dependencies)
xhtml2pdf==0.2.17
